+++ apps/portal/components/typeformApplicationSystem/index.tsx
@@ -1,0 +1,4 @@
+import TypeformEditForm from 'components/typeformApplicationSystem/TypeformEditForm';
+import TypeformCreateForm from 'components/typeformApplicationSystem/TypeformCreateForm';
+
+export { TypeformEditForm, TypeformCreateForm };
