+++ apps/portal/pages/admin/vanity/index.tsx
@@ -1,6 +1,6 @@
 import { useSession } from 'next-auth/react';
 import LoadingComponent from 'components/LoadingComponent';
-import VanityForm from 'components/VanityForm';
+import VanityForm from 'components/vanity/VanityForm';
 
 export default function CreateVanityLinkPage() {
   const { status } = useSession({ required: true });
